Split up the axis 360 api code into multiple files
