+++ bin/informational/axis-raw-bibliographic
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 from xml.dom import minidom
 
-from palace.manager.api.axis import Axis360API
+from palace.manager.api.axis.api import Axis360API
 from palace.manager.scripts.input import IdentifierInputScript
 from palace.manager.sqlalchemy.model.collection import Collection
