+++ src/palace/manager/api/axis/parser.py
@@ -1,39 +1,24 @@
 from __future__ import annotations
 
-import base64
 import datetime
 import html
 import json
 import re
-import ssl
-import urllib
 from abc import ABC, abstractmethod
-from collections.abc import Generator, Mapping, Sequence
-from dataclasses import dataclass
-from datetime import timedelta
-from typing import Any, Generic, Literal, Optional, TypeVar, Union, cast
-from urllib.parse import urlparse
+from collections.abc import Mapping
+from typing import TYPE_CHECKING, Any, Generic, Literal, Optional, TypeVar, Union, cast
 
-import certifi
-from flask_babel import lazy_gettext as _
 from lxml import etree
 from lxml.etree import _Element, _ElementTree
-from pydantic import field_validator
-from requests import Response as RequestsResponse
-from sqlalchemy.orm import Session
 
-from palace.manager.api.admin.validator import Validator
-from palace.manager.api.circulation import (
-    BaseCirculationAPI,
-    BaseCirculationApiSettings,
-    BaseCirculationLoanSettings,
-    CirculationInternalFormatsMixin,
-    Fulfillment,
-    HoldInfo,
-    LoanInfo,
-    PatronActivityCirculationAPI,
-    UrlFulfillment,
+from palace.manager.api.axis.constants import Axis360APIConstants
+from palace.manager.api.axis.fulfillment import (
+    Axis360AcsFulfillment,
+    Axis360Fulfillment,
 )
+from palace.manager.api.axis.loan_info import AxisLoanInfo
+from palace.manager.api.axis.manifest import AxisNowManifest
+from palace.manager.api.circulation import Fulfillment, HoldInfo
 from palace.manager.api.circulation_exceptions import (
     AlreadyCheckedOut,
     AlreadyOnHold,
@@ -52,9 +37,7 @@ from palace.manager.api.circulation_exceptions import (
     PatronLoanLimitReached,
     RemoteInitiatedServerError,
 )
-from palace.manager.api.selftest import HasCollectionSelfTests, SelfTestResult
 from palace.manager.api.web_publication_manifest import FindawayManifest, SpineItem
-from palace.manager.core.config import CannotLoadConfiguration
 from palace.manager.core.exceptions import IntegrationException
 from palace.manager.data_layer.bibliographic import BibliographicData
 from palace.manager.data_layer.circulation import CirculationData
@@ -62,696 +45,21 @@ from palace.manager.data_layer.contributor import ContributorData
 from palace.manager.data_layer.format import FormatData
 from palace.manager.data_layer.identifier import IdentifierData
 from palace.manager.data_layer.link import LinkData
-from palace.manager.data_layer.policy.replacement import ReplacementPolicy
 from palace.manager.data_layer.subject import SubjectData
-from palace.manager.integration.settings import (
-    ConfigurationFormItem,
-    ConfigurationFormItemType,
-    FormField,
-)
 from palace.manager.sqlalchemy.constants import LinkRelations, MediaTypes
 from palace.manager.sqlalchemy.model.classification import Classification, Subject
-from palace.manager.sqlalchemy.model.collection import Collection
 from palace.manager.sqlalchemy.model.contributor import Contributor
 from palace.manager.sqlalchemy.model.datasource import DataSource
 from palace.manager.sqlalchemy.model.edition import Edition
 from palace.manager.sqlalchemy.model.identifier import Identifier
-from palace.manager.sqlalchemy.model.licensing import (
-    DeliveryMechanism,
-    LicensePool,
-    LicensePoolDeliveryMechanism,
-)
-from palace.manager.sqlalchemy.model.patron import Patron
+from palace.manager.sqlalchemy.model.licensing import DeliveryMechanism, LicensePool
 from palace.manager.sqlalchemy.model.resource import Hyperlink, Representation
-from palace.manager.util.datetime_helpers import strptime_utc, utc_now
-from palace.manager.util.flask_util import Response
-from palace.manager.util.http import HTTP, RemoteIntegrationException
+from palace.manager.util.datetime_helpers import strptime_utc
 from palace.manager.util.log import LoggerMixin
 from palace.manager.util.xmlparser import XMLProcessor
 
-
-class Axis360APIConstants:
-    VERIFY_SSL = "verify_certificate"
-    PRODUCTION_BASE_URL = "https://axis360api.baker-taylor.com/Services/VendorAPI/"
-    QA_BASE_URL = "http://axis360apiqa.baker-taylor.com/Services/VendorAPI/"
-    SERVER_NICKNAMES = {
-        "production": PRODUCTION_BASE_URL,
-        "qa": QA_BASE_URL,
-    }
-
-
-class Axis360Settings(BaseCirculationApiSettings):
-    username: str = FormField(
-        form=ConfigurationFormItem(label=_("Username"), required=True)
-    )
-    password: str = FormField(
-        form=ConfigurationFormItem(label=_("Password"), required=True)
-    )
-    external_account_id: str = FormField(
-        form=ConfigurationFormItem(
-            label=_("Library ID"),
-            required=True,
-        )
-    )
-    url: str = FormField(
-        default=Axis360APIConstants.PRODUCTION_BASE_URL,
-        form=ConfigurationFormItem(
-            label=_("Server"),
-            required=True,
-        ),
-    )
-    verify_certificate: bool | None = FormField(
-        default=True,
-        form=ConfigurationFormItem(
-            label=_("Verify SSL Certificate"),
-            description=_(
-                "This should always be True in production; though, it may need "
-                "to be set to False to use the Axis 360 QA Environment."
-            ),
-            type=ConfigurationFormItemType.SELECT,
-            options={
-                "True": _("True"),
-                "False": _("False"),
-            },
-        ),
-    )
-
-    @field_validator("url")
-    @classmethod
-    def _validate_url(cls, v: str) -> str:
-        # Validate if the url provided is valid http or a valid nickname
-        valid_names = list(Axis360APIConstants.SERVER_NICKNAMES.keys())
-        if not Validator._is_url(v, valid_names):
-            raise ValueError(
-                f"Server nickname must be one of {valid_names}, or an 'http[s]' URL."
-            )
-        return v
-
-
-class Axis360LibrarySettings(BaseCirculationLoanSettings):
-    pass
-
-
-class Axis360API(
-    PatronActivityCirculationAPI[Axis360Settings, Axis360LibrarySettings],
-    HasCollectionSelfTests,
-    CirculationInternalFormatsMixin,
-    Axis360APIConstants,
-):
-    SET_DELIVERY_MECHANISM_AT = BaseCirculationAPI.BORROW_STEP
-
-    DATE_FORMAT = "%m-%d-%Y %H:%M:%S"
-
-    access_token_endpoint = "accesstoken"
-    availability_endpoint = "availability/v2"
-    fulfillment_endpoint = "getfullfillmentInfo/v2"
-    audiobook_metadata_endpoint = "getaudiobookmetadata/v2"
-
-    # Create a lookup table between common DeliveryMechanism identifiers
-    # and Axis 360 format types.
-    epub = Representation.EPUB_MEDIA_TYPE
-    pdf = Representation.PDF_MEDIA_TYPE
-    adobe_drm = DeliveryMechanism.ADOBE_DRM
-    findaway_drm = DeliveryMechanism.FINDAWAY_DRM
-    no_drm = DeliveryMechanism.NO_DRM
-    axisnow_drm = DeliveryMechanism.AXISNOW_DRM
-
-    # The name Axis 360 gives to its web interface. We use it as the
-    # name for the underlying access control system.
-    AXISNOW = "AxisNow"
-    BLIO = "Blio"
-
-    delivery_mechanism_to_internal_format = {
-        (epub, no_drm): "ePub",
-        (epub, adobe_drm): "ePub",
-        (pdf, no_drm): "PDF",
-        (pdf, adobe_drm): "PDF",
-        (None, findaway_drm): "Acoustik",
-        (None, axisnow_drm): AXISNOW,
-    }
-
-    @classmethod
-    def settings_class(cls) -> type[Axis360Settings]:
-        return Axis360Settings
-
-    @classmethod
-    def library_settings_class(cls) -> type[Axis360LibrarySettings]:
-        return Axis360LibrarySettings
-
-    @classmethod
-    def label(cls) -> str:
-        return "Axis 360"
-
-    @classmethod
-    def description(cls) -> str:
-        return ""
-
-    def __init__(
-        self, _db: Session, collection: Collection, bearer_token: str | None = None
-    ) -> None:
-        super().__init__(_db, collection)
-        settings = self.settings
-        self.library_id = settings.external_account_id
-        self.username = settings.username
-        self.password = settings.password
-
-        # Convert the nickname for a server into an actual URL.
-        base_url = settings.url or self.PRODUCTION_BASE_URL
-        if base_url in self.SERVER_NICKNAMES:
-            base_url = self.SERVER_NICKNAMES[base_url]
-        if not base_url.endswith("/"):
-            base_url += "/"
-        self.base_url = base_url
-
-        if not self.library_id or not self.username or not self.password:
-            raise CannotLoadConfiguration("Axis 360 configuration is incomplete.")
-
-        self._cached_bearer_token: str | None = bearer_token
-        self.verify_certificate: bool = (
-            settings.verify_certificate
-            if settings.verify_certificate is not None
-            else True
-        )
-
-    @property
-    def source(self) -> DataSource:
-        return DataSource.lookup(self._db, DataSource.AXIS_360, autocreate=True)
-
-    @property
-    def authorization_headers(self) -> dict[str, str]:
-        authorization = ":".join([self.username, self.password, self.library_id])
-        authorization_encoded = authorization.encode("utf_16_le")
-        authorization_b64 = base64.standard_b64encode(authorization_encoded).decode(
-            "utf-8"
-        )
-        return dict(Authorization="Basic " + authorization_b64)
-
-    def _run_self_tests(self, _db: Session) -> Generator[SelfTestResult]:
-        result = self.run_test("Refreshing bearer token", self._refresh_bearer_token)
-        yield result
-        if not result.success:
-            # If we can't get a bearer token, there's no point running
-            # the rest of the tests.
-            return
-
-        def _count_events() -> str:
-            now = utc_now()
-            five_minutes_ago = now - timedelta(minutes=5)
-            count = len(list(self.recent_activity(since=five_minutes_ago)))
-            return "Found %d event(s)" % count
-
-        yield self.run_test(
-            "Asking for circulation events for the last five minutes", _count_events
-        )
-
-        if self.collection is None:
-            raise ValueError("Collection is None")
-
-        for library_result in self.default_patrons(self.collection):
-            if isinstance(library_result, SelfTestResult):
-                yield library_result
-                continue
-            library, patron, pin = library_result
-
-            def _count_activity() -> str:
-                result = self.patron_activity(patron, pin)
-                return "Found %d loans/holds" % len(result)
-
-            yield self.run_test(
-                "Checking activity for test patron for library %s" % library.name,
-                _count_activity,
-            )
-
-        # Run the tests defined by HasCollectionSelfTests
-        for result in super()._run_self_tests(_db):
-            yield result
-
-    def _refresh_bearer_token(self) -> str:
-        url = self.base_url + self.access_token_endpoint
-        headers = self.authorization_headers
-        response = self._make_request(
-            url, "post", headers, allowed_response_codes=[200]
-        )
-        return self.parse_token(response.content)
-
-    def bearer_token(self) -> str:
-        if not self._cached_bearer_token:
-            self._cached_bearer_token = self._refresh_bearer_token()
-        return self._cached_bearer_token
-
-    def request(
-        self,
-        url: str,
-        method: str = "get",
-        extra_headers: dict[str, str] | None = None,
-        data: Mapping[str, Any] | None = None,
-        params: Mapping[str, Any] | None = None,
-        request_retried: bool = False,
-        **kwargs: Any,
-    ) -> RequestsResponse:
-        """Make an HTTP request, acquiring/refreshing a bearer token
-        if necessary.
-        """
-        if not extra_headers:
-            extra_headers = {}
-        headers = dict(extra_headers)
-        headers["Authorization"] = "Bearer " + self.bearer_token()
-        headers["Library"] = self.library_id
-        response = self._make_request(
-            url=url,
-            method=method,
-            headers=headers,
-            data=data,
-            params=params,
-            **kwargs,
-        )
-        if response.status_code == 401 and not request_retried:
-            parsed = StatusResponseParser().process_first(response.content)
-            if parsed is None or parsed[0] in [1001, 1002]:
-                # The token is probably expired. Get a new token and try again.
-                # Axis 360's status codes mean:
-                #   1001: Invalid token
-                #   1002: Token expired
-                self._cached_bearer_token = None
-                return self.request(
-                    url=url,
-                    method=method,
-                    extra_headers=extra_headers,
-                    data=data,
-                    params=params,
-                    request_retried=True,
-                    **kwargs,
-                )
-
-        return response
-
-    def availability(
-        self,
-        patron_id: str | None = None,
-        since: datetime.datetime | None = None,
-        title_ids: list[str] | None = None,
-    ) -> RequestsResponse:
-        url = self.base_url + self.availability_endpoint
-        args = dict()
-        if since:
-            since_str = since.strftime(self.DATE_FORMAT)
-            args["updatedDate"] = since_str
-        if patron_id:
-            args["patronId"] = patron_id
-        if title_ids:
-            args["titleIds"] = ",".join(title_ids)
-        response = self.request(url, params=args, timeout=None)
-        return response
-
-    def get_fulfillment_info(self, transaction_id: str) -> RequestsResponse:
-        """Make a call to the getFulfillmentInfoAPI."""
-        url = self.base_url + self.fulfillment_endpoint
-        params = dict(TransactionID=transaction_id)
-        # We set an explicit timeout because this request can take a long time and
-        # the default was too short. Ideally B&T would fix this on their end, but
-        # in the meantime we need to work around it.
-        # TODO: Revisit this timeout. Hopefully B&T will fix the performance
-        #   of this endpoint and we can remove this. We should be able to query
-        #   our logs to see how long these requests are taking.
-        return self.request(url, "POST", params=params, timeout=15)
-
-    def get_audiobook_metadata(self, findaway_content_id: str) -> RequestsResponse:
-        """Make a call to the getaudiobookmetadata endpoint."""
-        base_url = self.base_url
-        url = base_url + self.audiobook_metadata_endpoint
-        params = dict(fndcontentid=findaway_content_id)
-        # We set an explicit timeout because this request can take a long time and
-        # the default was too short. Ideally B&T would fix this on their end, but
-        # in the meantime we need to work around it.
-        # TODO: Revisit this timeout. Hopefully B&T will fix the performance
-        #   of this endpoint and we can remove this. We should be able to query
-        #   our logs to see how long these requests are taking.
-        response = self.request(url, "POST", params=params, timeout=15)
-        return response
-
-    def checkin(self, patron: Patron, pin: str, licensepool: LicensePool) -> None:
-        """Return a book early.
-
-        :param patron: The Patron who wants to return their book.
-        :param pin: Not used.
-        :param licensepool: LicensePool for the book to be returned.
-        :raise CirculationException: If the API can't carry out the operation.
-        :raise RemoteInitiatedServerError: If the API is down.
-        """
-        title_id = licensepool.identifier.identifier
-        patron_id = patron.authorization_identifier
-        response = self._checkin(title_id, patron_id)
-        try:
-            CheckinResponseParser().process_first(response.content)
-        except etree.XMLSyntaxError:
-            raise RemoteInitiatedServerError(response.text, self.label())
-
-    def _checkin(self, title_id: str | None, patron_id: str | None) -> RequestsResponse:
-        """Make a request to the EarlyCheckInTitle endpoint."""
-        if title_id is None:
-            self.log.warning(
-                f"Calling _checkin with title_id None. This is likely a bug. Patron_id: {patron_id}."
-            )
-            title_id = ""
-
-        if patron_id is None:
-            self.log.warning(
-                f"Calling _checkin with patron_id None. This is likely a bug. Title_id: {title_id}."
-            )
-            patron_id = ""
-
-        url = self.base_url + "EarlyCheckInTitle/v3?itemID={}&patronID={}".format(
-            urllib.parse.quote(title_id),
-            urllib.parse.quote(patron_id),
-        )
-        return self.request(url, method="GET", verbose=True)
-
-    def checkout(
-        self,
-        patron: Patron,
-        pin: str | None,
-        licensepool: LicensePool,
-        delivery_mechanism: LicensePoolDeliveryMechanism | None,
-    ) -> LoanInfo:
-        # Because we have SET_DELIVERY_MECHANISM_AT set to BORROW_STEP,
-        # delivery mechanism should always be set, but mypy doesn't know
-        # that, so we assert it here for type safety.
-        assert delivery_mechanism is not None
-
-        title_id = licensepool.identifier.identifier
-        patron_id = patron.authorization_identifier
-        response = self._checkout(
-            title_id, patron_id, self.internal_format(delivery_mechanism)
-        )
-
-        try:
-            response_text = response.text
-            self.log.info(
-                f"patron_id={patron_id} tried to checkout title_id={title_id}: "
-                f"response_code = {response.status_code}, response_content={response_text}"
-            )
-            expiration_date = CheckoutResponseParser().process_first(response_text)
-            return LoanInfo.from_license_pool(licensepool, end_date=expiration_date)
-        except etree.XMLSyntaxError:
-            raise RemoteInitiatedServerError(response.text, self.label())
-
-    def _checkout(
-        self, title_id: str | None, patron_id: str | None, internal_format: str
-    ) -> RequestsResponse:
-        url = self.base_url + "checkout/v2"
-        args = dict(titleId=title_id, patronId=patron_id, format=internal_format)
-        self.log.info(
-            f"patron_id={patron_id} about to checkout title_id={title_id}, using format={internal_format} "
-            f"posting to url={url}"
-        )
-        response = self.request(url, data=args, method="POST")
-
-        return response
-
-    def fulfill(
-        self,
-        patron: Patron,
-        pin: str,
-        licensepool: LicensePool,
-        delivery_mechanism: LicensePoolDeliveryMechanism,
-    ) -> Fulfillment:
-        """Fulfill a patron's request for a specific book."""
-        identifier = licensepool.identifier
-        # This should include only one 'activity'.
-        internal_format = self.internal_format(delivery_mechanism)
-        log_messages: list[str] = [
-            f"arguments for patron_activity method: "
-            f"patron.id={patron.id},"
-            f"internal_format={internal_format}, "
-            f"licensepool.identifier={identifier}, "
-            f"patron_id={patron.id}"
-        ]
-        activities = self.patron_activity(
-            patron, pin, licensepool.identifier, internal_format, log_messages
-        )
-
-        log_messages.append(
-            f"Patron activities returned from patron_activity method: {activities}"
-        )
-
-        for loan in activities:
-            if not isinstance(loan, AxisLoanInfo):
-                continue
-            if not (
-                loan.identifier_type == identifier.type
-                and loan.identifier == identifier.identifier
-            ):
-                continue
-            # We've found the remote loan corresponding to this
-            # license pool.
-            fulfillment = loan.fulfillment
-            if not fulfillment or not isinstance(fulfillment, Fulfillment):
-                raise CannotFulfill()
-            return fulfillment
-        # If we made it to this point, the patron does not have this
-        # book checked out.
-        log_messages.insert(
-            0,
-            "Unable to fulfill because there is no active loan. See info statements below for details:",
-        )
-        self.log.error("\n  ".join(log_messages))
-        raise NoActiveLoan()
-
-    def place_hold(
-        self,
-        patron: Patron,
-        pin: str | None,
-        licensepool: LicensePool,
-        hold_notification_email: str | None,
-    ) -> HoldInfo:
-        if not hold_notification_email:
-            hold_notification_email = self.default_notification_email_address(
-                patron, pin
-            )
-
-        url = self.base_url + "addtoHold/v2"
-        identifier = licensepool.identifier
-        title_id = identifier.identifier
-        patron_id = patron.authorization_identifier
-        params = dict(
-            titleId=title_id, patronId=patron_id, email=hold_notification_email
-        )
-        response = self.request(url, params=params)
-        hold_position = HoldResponseParser().process_first(response.content)
-        hold_info = HoldInfo.from_license_pool(
-            licensepool,
-            start_date=utc_now(),
-            hold_position=hold_position,
-        )
-        return hold_info
-
-    def release_hold(self, patron: Patron, pin: str, licensepool: LicensePool) -> None:
-        url = self.base_url + "removeHold/v2"
-        identifier = licensepool.identifier
-        title_id = identifier.identifier
-        patron_id = patron.authorization_identifier
-        params = dict(titleId=title_id, patronId=patron_id)
-        response = self.request(url, params=params)
-        try:
-            HoldReleaseResponseParser().process_first(response.content)
-        except NotOnHold:
-            # Fine, it wasn't on hold and now it's still not on hold.
-            pass
-        # If we didn't raise an exception, we're fine.
-        return None
-
-    def patron_activity(
-        self,
-        patron: Patron,
-        pin: str | None,
-        identifier: Identifier | None = None,
-        internal_format: str | None = None,
-        log_messages: list[str] | None = None,
-    ) -> list[AxisLoanInfo | HoldInfo]:
-        if identifier:
-            assert identifier.identifier is not None
-            title_ids = [identifier.identifier]
-        else:
-            title_ids = None
-
-        availability = self.availability(
-            patron_id=patron.authorization_identifier, title_ids=title_ids
-        )
-
-        availability_content_str = availability.text
-        if log_messages:
-            log_messages.append(
-                f"arguments to availability call: title_ids={title_ids}"
-            )
-            log_messages.append(
-                f"response to availability call: status={availability.status_code}, content={availability_content_str}"
-            )
-        loan_info_list = list(
-            AvailabilityResponseParser(self, internal_format).process_all(
-                availability_content_str
-            )
-        )
-
-        return loan_info_list
-
-    def update_availability(self, licensepool: LicensePool) -> None:
-        """Update the availability information for a single LicensePool.
-
-        Part of the CirculationAPI interface.
-        """
-        self.update_licensepools_for_identifiers([licensepool.identifier])
-
-    def update_licensepools_for_identifiers(
-        self, identifiers: list[Identifier]
-    ) -> None:
-        """Update availability and bibliographic information for
-        a list of books.
-
-        If the book has never been seen before, a new LicensePool
-        will be created for the book.
-
-        The book's LicensePool will be updated with current
-        circulation information.
-        """
-        remainder = set(identifiers)
-        for bibliographic, availability in self._fetch_remote_availability(identifiers):
-            edition, ignore1, license_pool, ignore2 = self.update_book(bibliographic)
-            identifier = license_pool.identifier
-            if identifier in remainder:
-                remainder.remove(identifier)
-
-        # We asked Axis about n books. It sent us n-k responses. Those
-        # k books are the identifiers in `remainder`. These books have
-        # been removed from the collection without us being notified.
-        for removed_identifier in remainder:
-            self._reap(removed_identifier)
-
-    def update_book(
-        self,
-        bibliographic: BibliographicData,
-    ) -> tuple[Edition, bool, LicensePool, bool]:
-        """Create or update a single book based on bibliographic
-        and availability data from the Axis 360 API.
-
-        :param bibliographic: A BibliographicData object containing
-            bibliographic and circulation (ie availability) data about this title
-        """
-        # The axis the bibliographic metadata always includes the circulation data
-        assert bibliographic.circulation
-
-        license_pool, new_license_pool = bibliographic.circulation.license_pool(
-            self._db, self.collection
-        )
-
-        edition, new_edition = bibliographic.edition(self._db)
-        license_pool.presentation_edition = edition
-        policy = ReplacementPolicy(
-            identifiers=False,
-            subjects=True,
-            contributions=True,
-            formats=True,
-            links=True,
-        )
-
-        bibliographic.apply(self._db, edition, self.collection, replace=policy)
-        return edition, new_edition, license_pool, new_license_pool
-
-    def _fetch_remote_availability(
-        self, identifiers: list[Identifier]
-    ) -> Generator[tuple[BibliographicData, CirculationData]]:
-        """Retrieve availability information for the specified identifiers.
-
-        :yield: A stream of (BibliographicData, CirculationData) 2-tuples.
-        """
-        identifier_strings = self.create_identifier_strings(identifiers)
-        return self.availability_by_title_ids(title_ids=identifier_strings)
-
-    def _reap(self, identifier: Identifier) -> None:
-        """Update our local circulation information to reflect the fact that
-        the identified book has been removed from the remote
-        collection.
-        """
-        collection = self.collection
-        pool = identifier.licensed_through_collection(collection)
-        if not pool:
-            self.log.warning(
-                "Was about to reap %r but no local license pool in this collection.",
-                identifier,
-            )
-            return
-        if pool.licenses_owned == 0:
-            # Already reaped.
-            return
-        self.log.info("Reaping %r", identifier)
-
-        availability = CirculationData(
-            data_source_name=pool.data_source.name,
-            primary_identifier_data=IdentifierData.from_identifier(identifier),
-            licenses_owned=0,
-            licenses_available=0,
-            licenses_reserved=0,
-            patrons_in_hold_queue=0,
-        )
-        availability.apply(
-            self._db, collection, ReplacementPolicy.from_license_source(self._db)
-        )
-
-    def recent_activity(
-        self, since: datetime.datetime
-    ) -> Generator[tuple[BibliographicData, CirculationData]]:
-        """Find books that have had recent activity.
-
-        :yield: A sequence of (BibliographicData, CirculationData) 2-tuples
-        """
-        availability = self.availability(since=since)
-        content = availability.content
-        yield from BibliographicParser().process_all(content)
-
-    def availability_by_title_ids(
-        self,
-        title_ids: list[str],
-    ) -> Generator[tuple[BibliographicData, CirculationData]]:
-        """Find title availability for a list of titles
-        :yield: A sequence of (BibliographicData, CirculationData) 2-tuples
-        """
-        availability = self.availability(title_ids=title_ids)
-        content = availability.content
-        yield from BibliographicParser().process_all(content)
-
-    @classmethod
-    def create_identifier_strings(
-        cls, identifiers: Sequence[Identifier | str]
-    ) -> list[str]:
-        identifier_strings = []
-        for i in identifiers:
-            if isinstance(i, Identifier):
-                assert i.identifier is not None
-                value = i.identifier
-            else:
-                value = i
-            identifier_strings.append(value)
-
-        return identifier_strings
-
-    @classmethod
-    def parse_token(cls, token: bytes) -> str:
-        data = json.loads(token)
-        return data["access_token"]  # type: ignore[no-any-return]
-
-    def _make_request(
-        self,
-        url: str,
-        method: str,
-        headers: Mapping[str, str],
-        data: Mapping[str, Any] | None = None,
-        params: Mapping[str, Any] | None = None,
-        **kwargs: Any,
-    ) -> RequestsResponse:
-        """Actually make an HTTP request."""
-        return HTTP.request_with_timeout(
-            method, url, headers=headers, data=data, params=params, **kwargs
-        )
-
+if TYPE_CHECKING:
+    from palace.manager.api.axis.api import Axis360API
 
 T = TypeVar("T")
 
@@ -828,9 +136,9 @@ class BibliographicParser(
     Axis360Parser[tuple[BibliographicData, CirculationData]], LoggerMixin
 ):
     DELIVERY_DATA_FOR_AXIS_FORMAT = {
-        Axis360API.BLIO: None,  # Legacy format, handled the same way as AxisNow
+        Axis360APIConstants.BLIO: None,  # Legacy format, handled the same way as AxisNow
         "Acoustik": (None, DeliveryMechanism.FINDAWAY_DRM),  # Audiobooks
-        Axis360API.AXISNOW: None,  # Handled specially, for ebooks only.
+        Axis360APIConstants.AXISNOW: None,  # Handled specially, for ebooks only.
         "ePub": (Representation.EPUB_MEDIA_TYPE, DeliveryMechanism.ADOBE_DRM),
         "PDF": (Representation.PDF_MEDIA_TYPE, DeliveryMechanism.ADOBE_DRM),
     }
@@ -1078,11 +386,11 @@ class BibliographicParser(
             informal_name = format_tag.text
             seen_formats.append(informal_name)
 
-            if informal_name == Axis360API.BLIO:
+            if informal_name == Axis360APIConstants.BLIO:
                 # We will be adding an AxisNow FormatData.
                 blio_seen = True
                 continue
-            elif informal_name == Axis360API.AXISNOW:
+            elif informal_name == Axis360APIConstants.AXISNOW:
                 # We will only be adding an AxisNow FormatData if this
                 # turns out to be an ebook.
                 axisnow_seen = True
@@ -1368,17 +676,6 @@ class HoldReleaseResponseParser(XMLResponseParser[Literal[True]]):
         # or we were successful.
         self.raise_exception_on_error(e, namespaces, {3109: NotOnHold})
         return True
-
-
-@dataclass(kw_only=True)
-class AxisLoanInfo(LoanInfo):
-    """
-    An extension of the normal LoanInfo dataclass that includes some Axis 360-specific
-    information, since the Axis 360 API uses this object to get information about
-    loan fulfillment in addition to checkout.
-    """
-
-    fulfillment: Fulfillment | None
 
 
 class AvailabilityResponseParser(XMLResponseParser[Union[AxisLoanInfo, HoldInfo]]):
@@ -1707,166 +1004,3 @@ class AudiobookMetadataParser(
         part_number = int(part.get("fndpart", 0))
         sequence = int(part.get("fndsequence", 0))
         return SpineItem(title, duration, part_number, sequence)
-
-
-class AxisNowManifest:
-    """A simple media type for conveying an entry point into the AxisNow access control
-    system.
-    """
-
-    MEDIA_TYPE = DeliveryMechanism.AXISNOW_DRM
-
-    def __init__(self, book_vault_uuid: str, isbn: str):
-        """Constructor.
-
-        :param book_vault_uuid: The UUID of a Book Vault.
-        :param isbn: The ISBN of a book in that Book Vault.
-        """
-        self.book_vault_uuid = book_vault_uuid
-        self.isbn = isbn
-
-    def __str__(self) -> str:
-        data = dict(isbn=self.isbn, book_vault_uuid=self.book_vault_uuid)
-        return json.dumps(data, sort_keys=True)
-
-
-class Axis360Fulfillment(Fulfillment, LoggerMixin):
-    """An Axis 360-specific Fulfillment implementation for audiobooks
-    and books served through AxisNow.
-
-    We use these instead of normal Fulfillment objects because
-    putting all this information into Fulfillment would require
-    one or two extra HTTP requests, and there's often no need to make
-    those requests.
-    """
-
-    def __init__(
-        self,
-        api: Axis360API,
-        data_source_name: str,
-        identifier_type: str,
-        identifier: str,
-        key: str,
-    ):
-        """Constructor.
-
-        :param api: An Axis360API instance, in case the parsing of
-        a fulfillment document triggers additional API requests.
-
-        :param key: The transaction ID that will be used to fulfill
-        the request.
-        """
-        self.data_source_name = data_source_name
-        self.identifier_type = identifier_type
-        self.identifier = identifier
-        self.api = api
-        self.key = key
-
-        self.content_type: str | None = None
-        self.content: str | None = None
-
-    def license_pool(self, _db: Session) -> LicensePool:
-        """Find the LicensePool model object corresponding to this object."""
-        collection = self.api.collection
-        pool, is_new = LicensePool.for_foreign_id(
-            _db,
-            self.data_source_name,
-            self.identifier_type,
-            self.identifier,
-            collection=collection,
-        )
-        return pool
-
-    def do_fetch(self) -> tuple[str, str]:
-        _db = self.api._db
-        license_pool = self.license_pool(_db)
-        transaction_id = self.key
-        response = self.api.get_fulfillment_info(transaction_id)
-        parser = Axis360FulfillmentInfoResponseParser(self.api)
-        manifest, expires = parser.parse(response.content, license_pool=license_pool)
-        return str(manifest), manifest.MEDIA_TYPE
-
-    def response(self) -> Response:
-        if self.content is None:
-            self.content, self.content_type = self.do_fetch()
-        return Response(response=self.content, content_type=self.content_type)
-
-
-class Axis360AcsFulfillment(UrlFulfillment, LoggerMixin):
-    """This implements a Axis 360 specific Fulfillment for ACS content
-    served through AxisNow. The AxisNow API gives us a link that we can use
-    to get the ACSM file that we serve to the mobile apps.
-
-    This link resolves to a redirect, which resolves to the actual ACSM file.
-    The URL we are given in the redirect has a percent encoded query string
-    in it. The encoding used in this string has lower case characters in it
-    like "%3a" for :.
-
-    In versions of urllib3 > 1.24.3 the library normalizes the query string
-    before doing the actual request. In doing the normalization it follows the
-    recommendation of RFC 3986 and uppercases the percent encoded bytes.
-
-    This causes the Axis360 API to return an error from Adobe ACS:
-    ```
-    <error xmlns="http://ns.adobe.com/adept" data="E_URLLINK_AUTH
-    https://acsqa.digitalcontentcafe.com/fulfillment/URLLink.acsm"/>
-    ```
-    instead of the correct ACSM file.
-
-    Others have noted that this is a problem in the urllib3 github but they
-    do not seem interested in providing an option to override this behavior
-    and closed the ticket.
-    https://github.com/urllib3/urllib3/issues/1677
-
-    This Fulfillment implementation uses the built in Python urllib
-    implementation instead of requests (and urllib3) to make this request
-    to the Axis 360 API, sidestepping the problem, but taking a different
-    code path than most of our external HTTP requests.
-    """
-
-    def __init__(
-        self,
-        content_link: str,
-        verify: bool,
-        content_type: str = DeliveryMechanism.ADOBE_DRM,
-    ) -> None:
-        super().__init__(content_link, content_type)
-        self.verify = verify
-
-    def response(self) -> Response:
-        service_name = urlparse(str(self.content_link)).netloc
-        try:
-            if self.verify:
-                # Actually verify the ssl certificates
-                ssl_context = ssl.SSLContext(ssl.PROTOCOL_TLS)
-                ssl_context.verify_mode = ssl.CERT_REQUIRED
-                ssl_context.check_hostname = True
-                ssl_context.load_verify_locations(cafile=certifi.where())
-            else:
-                # Default context does no ssl verification
-                ssl_context = ssl.SSLContext()
-            req = urllib.request.Request(self.content_link)
-            with urllib.request.urlopen(
-                req, timeout=20, context=ssl_context
-            ) as response:
-                content = response.read()
-                status = response.status
-                headers = response.headers
-
-        # Mimic the behavior of the HTTP.request_with_timeout class and
-        # wrap the exceptions thrown by urllib and ssl by raising a RemoteIntegrationException
-        except urllib.error.HTTPError as e:
-            message = f"The server received a bad status code ({e.code}) while contacting {service_name}"
-            self.log.warning(message)
-            raise RemoteIntegrationException(service_name, message) from e
-        except TimeoutError as e:
-            message = f"Error connecting to {service_name}. Timeout occurred."
-            self.log.warning(message)
-            raise RemoteIntegrationException(service_name, message) from e
-        except (urllib.error.URLError, ssl.SSLError) as e:
-            reason = getattr(e, "reason", e.__class__.__name__)
-            message = f"Error connecting to {service_name}. {reason}."
-            self.log.warning(message)
-            raise RemoteIntegrationException(service_name, message) from e
-
-        return Response(response=content, status=status, headers=headers)
