+++ tests/manager/core/test_coverage.py
@@ -3,7 +3,7 @@ from typing import cast
 
 import pytest
 
-from palace.manager.api.axis import Axis360API
+from palace.manager.api.axis.api import Axis360API
 from palace.manager.api.overdrive.api import OverdriveAPI
 from palace.manager.core.coverage import (
     BaseCoverageProvider,
